+++ packages/orq-rc/RELEASES.md
@@ -8489,3 +8489,13 @@ Based on:
 - [typescript v4.8.0-rc.42] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.42 - packages/orq-rc
+
+## 2026-04-21 21:06:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.43] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.43 - packages/orq-rc